requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dataclasses-json==0.6.7
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.352
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
lxml==5.3.0
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==0.27.8
packaging==23.2
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
python-docx==1.1.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
yarl==1.11.1
